Exported booking defined data visualization into a separate component
